chore(dashboard): create a dashboard module
